requirements.txt
# KI-Mitarbeiter System Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# HTTP Client (for GitHub API and Kimi API)
httpx==0.26.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1
numpy==1.26.4

# Cache
redis==5.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Optional: For production
# gunicorn==21.2.0
